@@ -15,6 +15,7 @@ dependencies:
   carrier_info: ^2.0.5
   cupertino_icons: ^1.0.5
   cupertino_lists: ^1.0.1
+  cupertino_onboarding: ^1.0.1
   dart_ping: ^6.1.2
   dart_ping_ios: ^1.1.0
   dynamic_color: ^1.4.0
@@ -28,12 +29,12 @@ dependencies:
     sdk: flutter
   flutter_bloc: ^8.0.1
   flutter_platform_widgets: ^2.0.0
-  flutter_svg: ^1.0.3
+  flutter_svg: ^1.1.1
   font_awesome_flutter: ^10.1.0
   google_fonts: ^3.0.1
   google_maps_flutter: ^2.1.8
   google_mobile_ads: ^1.3.0
-  hive: ^2.2.2
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
   http: ^0.13.4
   hydrated_bloc: ^8.1.0
@@ -43,13 +44,13 @@ dependencies:
   json_annotation: ^4.5.0
   json_serializable: ^6.2.0
   lan_scanner: ^3.4.1
-  modal_bottom_sheet: ^2.0.1
+  modal_bottom_sheet: ^2.1.0
   network_info_plus: ^2.1.3
   package_info_plus: ^1.4.2
   path_provider: ^2.0.11
   permission_handler: ^10.0.0
-  sentry_flutter: ^6.5.1
-  url_launcher: ^6.1.3
+  sentry_flutter: ^6.6.1
+  url_launcher: ^6.1.4
   wake_on_lan: ^3.0.0
 
 dev_dependencies:
